requirements.txt
fastapi==0.116.1
fitz==0.0.1.dev2
langchain==0.3.26
langchain_core==0.3.68
langchain_neo4j==0.4.0
langchain_openai==0.3.28
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.4
